build 되도록 Long -> Integer 로 변경 및 dependency 추가

+++ src/main/java/zerobase/matching/chat/entity/ChatRoom.java
@@ -17,7 +17,7 @@ public class ChatRoom {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "ROOM_ID")
-    private Long chatroomId;
+    private int chatroomId;
 
     @Column(name = "TITLE")
     private String title;
